Create c.php
This just needs a proper decompressor
The first bit is the bit to use
the 2nd is the type of order
  - if same, all bits in a row are the same
  - if opposite, all bits flip flop the distance
The flipping bits are going to be managed
by the opposite of the 1st bit between 2 bits which offer $assertion%4
which is of course right-shifted off the number.
Yep, that's it. Figure you have the manpower to shrink the problem here
it's worth it I think.